move dma driver code from register definitions to dma drivers
